Added an About Us Screen and updated the Details Screen
Added an About Us screen, a prototype for now until I can get some better styling.

I tried to make the parameters of the Details take in the string values for gym name and gym leader, but it still messy with the image url. Will cause a crash.  Working on it.